feat: add lock capabilities to settings
* Implement locked settings exception handling in repository
* Refactor code
* Remove unnecessary code
The removal is necessary because is no longer needed because the update and delete logic is handle directly in the `repository` and not in the model. Also, because we update one or more than one setting at the time, we can quickly trigger the N+1 problem, or trying to rehydrate the model to check if is locked is not necessary. In the case of a deletion we just run a single extra query, and for updating, we get the information directly from the current query, so this does not add any heavy calculation to the code.
* Fix code style issues
* Enhance LockedSettingException for pluralized messages and update tests
* Refactor expectations to use a enhance version
* Refactor test to use byGroup method for consistency and fix typo

+++ src/Repositories/EloquentRepository.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Coyotito\LaravelSettings\Repositories;
 
+use Coyotito\LaravelSettings\Models\Exceptions\LockedSettingException;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Arr;
 use RuntimeException;
@@ -73,7 +74,12 @@ class EloquentRepository implements Contracts\Repository
 
         $settingNames = array_keys($setting);
 
-        $presentSettings = $this->withGroup()->whereIn('name', $settingNames)->pluck('name');
+        $presentSettings = $this->withGroup()->whereIn('name', $settingNames)->pluck('locked', 'name');
+        $locked = $presentSettings->filter(fn (bool $locked): bool => $locked);
+
+        if ($locked->isNotEmpty()) {
+            throw new LockedSettingException($locked->keys()->all());
+        }
 
         if ($presentSettings->isEmpty()) {
             return;
@@ -81,7 +87,7 @@ class EloquentRepository implements Contracts\Repository
 
         $now = now();
 
-        $data = $presentSettings->map(function (string $name) use ($setting, $now): array {
+        $data = $presentSettings->keys()->map(function (string $name) use ($setting, $now): array {
             return [
                 'name' => $name,
                 'group' => $this->group(),
@@ -128,6 +134,12 @@ class EloquentRepository implements Contracts\Repository
 
         if (empty($setting)) {
             return 0;
+        }
+
+        $locked = $this->withGroup()->where('locked', true)->whereIn('name', $setting)->pluck('name');
+
+        if ($locked->isNotEmpty()) {
+            throw new LockedSettingException($locked->all());
         }
 
         return (int) $this->withGroup()->whereIn('name', $setting)->delete();
